Add files via upload
This is my STA 309 Midterm 2 final submission. This project takes earthquake data and gives us the freedom to make our own dashboard. I decided to research earthquake detection history and see how different technology affects detection over the year. The Global Seismographic Network was implemented in 1984. GSN is a permanent digital network of seismometers that measure the vibrations of the world. I focused on 1974-1994 to see the earthquake trends before and immediately after the implementation of GSN. I found that GSN created more consistent earthquake predictions with less variance.